♻️Refactor - 오픈세미나 멤버 기본이미지로 변경
